refactor: move TranslatedModelSerializer to linkedevents.serializers
